Contrast analyses for dependent samples

This function allows to directly compare two contrasts for dependent samples by standardizing the contrast weights. Please note that sample sizes must be equal between within-subject groups, so no missings are allowed

compare_dependent(nGroup, lambda1, lambda2, dat, testvalue = 0)

Arguments

nGroup
Number of dependent / within-subject groups

lambda1
A vector of contrast weights for Hypothesis 1

lambda2
A vector of contrast weights for Hypothesis 2

dat
A matrix or dataframe with nGroup columns; each row contains values for one respondent; each column contains values of the dependent variable in the respective within-subject group

testvalue
value under the Null hypothesis; if not specified, it is fixed to 0.

Source

Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)

Value

A list with following entries:

Results

Results of the contrast analysis

Contrast Weights

Standardized contrast weights of the two original contrasts and difference between standardized contrast weights

Note

A test favoring Hypothesis 1 is performed, hence a positive t-value indicate that the contrast weights contained in lambda1 fit the data better than the contrast weights contained in lambda2, and vice versa for a negative t-value.

Examples

# load presidents dataset
presidents <- data.frame(Qtr1=presidents[seq(1, length(presidents), 4)],
               Qtr2=presidents[seq(2, length(presidents), 4)],
               Qtr3=presidents[seq(3, length(presidents), 4)],
               Qtr4=presidents[seq(4, length(presidents), 4)])
presidents <- na.omit(presidents)

# define lambda weights
lambda1 <- c(1, 0, 0, -1)   # H1: decrease in approval ratings with stagnation over warmer months
lambda2 <- c(3, 1, -1, -3)  # H2: linear decrease in  approval ratings

# perform contrast analysis
compare_dependent(nGroup=4, lambda1, lambda2, presidents)

#> $results
#>                 F estimate      t      p      g
#> Contrast 1 0.2958   0.3858 0.5439 0.5914 0.1067
#> 
#> $contrastWeights
#>   lambda1Std lambda2Std  lambdaDiff
#> 1   1.414214  1.3416408  0.07257278
#> 2   0.000000  0.4472136 -0.44721360
#> 3   0.000000 -0.4472136  0.44721360
#> 4  -1.414214 -1.3416408 -0.07257278
#>