Contrast analyses for independent samples

This function allows to directly compare two contrasts for independent samples by standardizing the contrast weights

compare_independent(n_group, lambda_preferred, lambda_competing, data)

Source

Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)

Arguments

n_group

Number of independent / between-subject groups

lambda_preferred

A vector of contrast weights for Hypothesis 1

lambda_competing

A vector of contrast weights for Hypothesis 2

data

A matrix or dataframe with two columns; each row contains values for one respondents; the first column contains the group indicator, the second column contains the dependent variable

Value

A list with following entries:

Results

Results of the contrast analysis

Contrast Weights

Standardized contrast weights of the two original contrasts and difference between standardized contrast weights

Note

A test favoring Hypothesis 1 is performed. This means that when the t-value is positive, the contrast weights contained in lambda_preferred reflect the data better than the contrast weights contained in lambda_competing. When the t-value is negative, the contrast weights contained in lambda_competing reflect the data better, and Hypothesis 2 is favored over Hypothesis 1.

Examples

# load iris dataset
data("iris")
iris <- iris[,c("Species", "Petal.Length")]

# define lambda weights
lambda_preferred <- c(-1, 0, 1)  # H1: Iris versicolor is distinct to Iris setosa and Iris virginica
lambda_competing <- c(-2, 1, 1)  # H2: Iris versicolor is more similar to Iris virginica

# perform contrast analysis
compare_independent(n_group=3, lambda_preferred, lambda_competing, iris)
#> Please note that each contrast is tested separately, the contrast are not tested jointly!
#> * When the contrasts are orthogonal and the samples size is equal in all groups, the results from contrasts tested separately is equal to the results from contrasts
#> tested together.
#> * However, the results may differ when non-orthogonal contrasts are tested separately.
#> * You may want to switch to an alternative package to conduct a joint test of all contrasts
#> $results
#>                SS  df contrast_estimate t_value F_value p_value r_effect_size
#> Contrast 1 1.1958 147            0.1387  2.5411  6.4574  0.0121        0.0507
#>            r_alerting r2_alerting r_contrast
#> Contrast 1     0.0523      0.0027     0.2051
#> 
#> $contrast_weights
#>   lambda_preferred_std lambda_competing_std lambda_diff
#> 1            -1.224745           -1.4142136   0.1894687
#> 2             0.000000            0.7071068  -0.7071068
#> 3             1.224745            0.7071068   0.5176381
#>