Contrast analyses for dependent samples

This function allows to perform contrast analyses for dependent samples. Please note that sample sizes must be equal between within-subject groups, so no missings are allowed

contrast_dependent(nGroup, lambda, dat, testvalue = 0)

Arguments

nGroup
Number of dependent / within-subject groups

lambda
A matrix of contrast weights with contrasts in rows and groups in columns

dat
A matrix or dataframe with nGroup columns; each row contains values for one respondent; each column contains values of the dependent variable in the respective within-subject group

testvalue
Value under the null hypothesis; if not specified, it is fixed to 0.

Source

Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)

Value

a dataframe with following entries for each of the contrasts:

SumsofSquares

Sums of Squares

df

Degrees of freedom for each contrast

F

F-values

contrast estimate

Contrast estimates

t

t-values

p

Two-tailed p-values

g

Effect size g, a standardized distance measure

Examples

# load presidents dataset
presidents <- data.frame(Qtr1=presidents[seq(1, length(presidents), 4)],
               Qtr2=presidents[seq(2, length(presidents), 4)],
               Qtr3=presidents[seq(3, length(presidents), 4)],
               Qtr4=presidents[seq(4, length(presidents), 4)])
presidents <- na.omit(presidents)
nGroup <- ncol(presidents)

# define lambda weights
lambda <- matrix(c(
    1, 0, 0, -1, # H1: decrease in approval ratings with stagnation over warmer months
    3, 1, -1, -3), # H2: linear decrease in  approval ratings
    ncol = nGroup,
    byrow=TRUE)

# perform contrast analysis
contrast_dependent(nGroup, lambda,presidents)

#>                  F df contrast.estimate      t      p      g
#> Contrast 1  9.6148  1            7.9231 3.1008 0.0047 0.6081
#> Contrast 2 10.4904  1           24.1923 3.2389 0.0034 0.6352