Contrast analyses for independent samples

This function allows to perform contrast analyses for independent samples

contrast_independent(nGroup, lambda, dat)

Arguments

nGroup
Number of independent / between-subject groups

lambda
A matrix of contrast weights with contrasts in rows and groups in columns

dat
A matrix or dataframe with two columns; each row contains values for one respondent; the first column contains the group indicator, the second column contains the dependent variable

Source

Rosenthal et al. (2000); Sedlmeier & Renkewitz (2013)

Value

A dataframe with following entries for each of the contrasts:

SumsofSquares

Sums of Squares

df

Degrees of freedom for each contrast

F

F-values

contrast estimate

Contrast estimates

t

t-values

p

Two-tailed p-values

rEffectSize

Correlation between the dependent variable and the contrast weights

rAlerting

Correlation between group means and contrast weights

r2Aalerting

Squared rAlerting; can be interpreted similar to a determination coefficient as a measured of explained variance by the contrast (SScontrast/SSbetween)

rContrast

sqrt(t^2/(t^2 + df)); useful for power analyses

Examples

# load iris dataset
data("iris")
iris <- iris[,c("Species", "Petal.Length")]
nGroup <- length(levels(iris$Species))

# define lambda weights
lambda <- matrix(c(
                -1,0,1, # H1: An increase from setosa over versicolor to virginica
                -2,1,1), # H2: Setosa has smaller Petal Length than versicolor and virginica
                ncol = nGroup,
                byrow=TRUE)
# perform contrast analysis
contrast_independent(nGroup, lambda, iris)

#>            SumsofSquares df        F contrast.estimate       t p rEffectSize
#> Contrast 1      418.2025  1 2258.262             4.090 47.5212 0      0.9490
#> Contrast 2      395.3712  1 2134.975             6.888 46.2058 0      0.9228
#>            rAlerting r2Alerting rContrast
#> Contrast 1    0.9781     0.9568    0.9690
#> Contrast 2    0.9511     0.9045    0.9673