Working memory data

This dataset originates from a study conducted at the University of Zurich in the area of visual working memory. This is a vast simplification of the original study for illustrative purposes. The experiment consisted of 324 trials per participant. They were presented with eather two, four, or eight colored squares (set size manipulation) and had to reproduce the color shortly afterwards. Then they were asked to judge their confidence on the given response.

working_memory

Format

a data frame with 6660 rows and 4 variables:

ID

the participant's id

Trial

trial indicator

SetSize

set size manipulation (within-subjects)

confidence

confidence rating

Details

For more details see: https://psyarxiv.com/xmf24/

For the entire dataset and analysis scripts see: https://osf.io/5mgzt/